% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput.R
\name{addAnnotations}
\alias{addAnnotations}
\title{Add annotations}
\usage{
addAnnotations(study, annotations, overwrite = FALSE)
}
\arguments{
\item{annotations}{The annotations used for the enrichment analyses. The input
is a nested list. The top-level list contains one entry per annotation
database, e.g. reactome. The names correspond to the name of each annotation
database. Each of these elements should be list of annotation terms. The
names of this sublist correspond to the name of the annotation term. Each of
the annotation terms should be a character vector of feature IDs.}
}
\description{
Add annotations
}
