% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addContrasts}
\alias{addContrasts}
\title{Add contrasts}
\usage{
addContrasts(study, contrasts, overwrite = FALSE)
}
\arguments{
\item{contrasts}{The contrasts tested by the model(s). A named character
vector. The names correspond to the name of each contrast. The elements
correspond to the description of each contrast.}
}
\description{
Add contrasts
}
