% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addEnrichments}
\alias{addEnrichments}
\title{Add enrichment results}
\usage{
addEnrichments(study, enrichments, overwrite = FALSE)
}
\arguments{
\item{enrichments}{The enrichment results from each model. The input is a
nested named list. The names of the list correspond to the model names.
Each list element should be a list of the tests tested. The names
correspond to the test names. Each list element should be another list of
annotation databases. The names correspond to the annotation databases.
Each of these elements should be a data frame with enrichment results. Each
table must contain the following columns: "termID", "description",
"nominal" (the nominal statistics), and "adjusted" (the statistics after
adjusting for multiple testing). Any additional columns are ignored.}
}
\description{
To reduce storage and compute time, it is highly recommended to first filter
the enrichment results for statistical signficance prior to adding them to
the study.
}
