% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addEnrichmentsLinkouts}
\alias{addEnrichmentsLinkouts}
\title{Add linkouts to external resources in the enrichments table}
\usage{
addEnrichmentsLinkouts(study, enrichmentsLinkouts)
}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}

\item{enrichmentsLinkouts}{The URL patterns that describe linkouts to
external resources (see Details below). The input object is a named list.
The names of the list correspond to the annotation names. Each element of
the list is a character vector of linkouts for that annotationID.}
}
\description{
You can provide additional information on the annotation terms in your study
by providing linkouts to external resources. These will be embedded directly
in the enrichments table.
}
\details{
For each linkout, the URL pattern you provide will be concatenated with the
value of the termID column. As an example, if you used the annotation
database \href{http://amigo.geneontology.org/}{AmiGO 2} for your enrichments
analysis, you can provide a linkout for each termID using the following
pattern:

\preformatted{go = "http://amigo.geneontology.org/amigo/term/"}

As another example, if you used the annotation database
\href{https://reactome.org/}{Reactome} for your enrichments analysis, you can
provide a linkout for each termID using the following pattern:

\preformatted{reactome = "https://reactome.org/content/detail/"}

Note that you can provide more than one linkout per termID.
}
\examples{
  study <- createStudy("example")
  enrichmentsLinkouts <- list(
    gobp = c("http://amigo.geneontology.org/amigo/term/",
             "https://www.ebi.ac.uk/QuickGO/term/"),
    reactome = "https://reactome.org/content/detail/"
  )
  study <- addEnrichmentsLinkouts(study, enrichmentsLinkouts)

}
\seealso{
\code{\link{addAnnotations}}, \code{\link{addEnrichments}}
}
