% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addFeatures}
\alias{addFeatures}
\title{Add feature metadata}
\usage{
addFeatures(study, features, overwrite = FALSE)
}
\arguments{
\item{features}{A table of metadata variables that describe the features in the
study. The table must contain the unique featureID used for the study. Also,
the object must inherit from the class data.frame.}
}
\description{
Add feature metadata
}
