% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addInferences}
\alias{addInferences}
\title{Add inference results}
\usage{
addInferences(study, inferences, overwrite = FALSE)
}
\arguments{
\item{inferences}{The inference results from each model. The input is a
nested named list. The names of the list correspond to the model names.
Each element in the list should be a list of data frames with inference
results, one for each contrast. The featureID column needs to be included
in each table.}
}
\description{
Add inference results
}
