% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addMetaAssays}
\alias{addMetaAssays}
\title{Add metaAssays}
\usage{
addMetaAssays(study, metaAssays, reset = FALSE)
}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}

\item{metaAssays}{The metaAssays from the study. The input object is a list
of data frames (one per model). The row names should correspond to the
metaFeatureIDs (second column of data frame added via
\code{\link{addMetaFeatures}}). The column names should correspond to the
sampleIDs (\code{\link{addSamples}}). The data frame should only contain
numeric values. To share a data frame across multiple models, use the
modelID "default".}

\item{reset}{Reset the data prior to adding the new data (default:
\code{FALSE}). The default is to add to or modify any previously added data
(if it exists). Setting \code{reset = TRUE} enables you to remove existing
data you no longer want to include in the study.}
}
\description{
Experimental. Add metaAssay measurements that map to the features in the
assays data.
}
\seealso{
\code{\link{addAssays}}, \code{\link{addMetaFeatures}}
}
