% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput.R
\name{addMetaFeatures}
\alias{addMetaFeatures}
\title{Add meta-feature metadata}
\usage{
addMetaFeatures(study, metaFeatures, overwrite = FALSE)
}
\arguments{
\item{metaFeatures}{A table of metadata variables that describe the
meta-features in the study. This is useful anytime there are metadata
variables that cannot be mapped 1:1 to your features. For example, a
peptide may be associated with multiple proteins. The table must contain
the unique featureID used for the study. Also, the object must inherit from
the class data.frame.}
}
\description{
Add meta-feature metadata
}
