% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{addOverlaps}
\alias{addOverlaps}
\title{Add overlaps between annotation gene sets}
\usage{
addOverlaps(study)
}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}
}
\value{
Returns the original \code{onStudy} object passed to the argument
  \code{study}, but modified to include the newly added data
}
\description{
The app's network view of the enrichments results requires pairwise overlap
metrics between all the terms of each annotation in order to draw the edges
between the nodes/terms. These overlaps are calculated automatically when
installing or exporting an OmicNavigator study. If you'd like, you can
manually calculate these pairwise overlaps by calling \code{addOverlaps}
prior to installing or exporting your study.
}
