% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addResults}
\alias{addResults}
\title{Add inference results}
\usage{
addResults(study, results)
}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}

\item{results}{The inference results from each model. The input is a
nested named list. The names of the list correspond to the model names.
Each element in the list should be a list of data frames with inference
results, one for each test. In each data frame, the featureID must be in
the first column, and all other columns must be numeric.}
}
\description{
Add inference results
}
