% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput.R
\name{addSamples}
\alias{addSamples}
\title{Add sample metadata}
\usage{
addSamples(study, samples, overwrite = FALSE)
}
\arguments{
\item{samples}{A table of metadata variables that describe the samples in the
study. The table must contain the unique sampleID used for the study. Also,
the object must inherit from the class data.frame.}
}
\description{
Add sample metadata
}
