% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTests}
\alias{addTests}
\title{Add tests}
\usage{
addTests(study, tests, overwrite = FALSE)
}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{tests}{The tests from the study. The input object can be a single data
frame or a list of data frames (one per model). The first column should
contain the unique ID for each test. The second column should contain a
description of the test.}

\item{overwrite}{Overwrite existing value}
}
\description{
Add tests
}
