% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basal.vs.ml}
\alias{basal.vs.ml}
\title{basal.vs.ml from Bioconductor workflow RNAseq123}
\format{
A data frame with 24 rows and 8 columns:
\describe{
  \item{ENTREZID}{Entrez ID of mouse gene}
  \item{SYMBOL}{Symbol of mouse gene}
  \item{TXCHROM}{Chromosome location of mouse gene}
  \item{logFC}{Log fold change}
  \item{AveExpr}{Average expression level of the gene across all samples}
  \item{t}{Moderated t-statistic}
  \item{P.Value}{p-value}
  \item{adj.P.Val}{Adjusted p-value}
}
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
basal.vs.ml
}
\description{
A subset of the object \code{basal.vs.ml} from Bioconductor workflow RNAseq123.
}
\examples{
  data("RNAseq123", package = "OmicAnalyzer")
  head(basal.vs.ml)
  str(basal.vs.ml)

}
\keyword{datasets}
