% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cam.BasalvsML}
\alias{cam.BasalvsML}
\title{cam.BasalvsML from Bioconductor workflow RNAseq123}
\format{
A data frame with 4 rows and 4 columns:
\describe{
  \item{NGenes}{Number of genes in each term}
  \item{Direction}{Direction of the enrichment}
  \item{PValue}{Nominal p-value}
  \item{FDR}{Multiple-testing adjusted p-value}
}
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
cam.BasalvsML
}
\description{
A subset of the object \code{cam.BasalvsML} from Bioconductor workflow RNAseq123.
}
\examples{
  data("RNAseq123", package = "OmicAnalyzer")
  head(cam.BasalvsML)
  str(cam.BasalvsML)

}
\keyword{datasets}
