% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportStudy}
\alias{exportStudy}
\title{Export a study}
\usage{
exportStudy(
  study,
  type = c("tarball", "package"),
  path = NULL,
  requireValid = TRUE
)
}
\arguments{
\item{study}{An OmicNavigator study}

\item{type}{Export study as a package tarball ("tarball") or as a package
directory ("package")}

\item{path}{Optional file path to save the object}

\item{requireValid}{Require that study is valid before exporting (via
\code{\link{validateStudy}})}
}
\value{
Invisibly returns the name of the tarball file ("tarball") or the
path to the package directory ("package")
}
\description{
Export a study
}
\seealso{
\code{\link{validateStudy}}
}
