% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{genes from Bioconductor workflow RNAseq123}
\format{
A data frame with 24 rows and 3 columns:
\describe{
  \item{ENTREZID}{Entrez ID of mouse gene}
  \item{SYMBOL}{Symbol of mouse gene}
  \item{TXCHROM}{Chromosome location of mouse gene}
}
}
\source{
\url{https://bioconductor.org/packages/release/workflows/html/RNAseq123.html}
}
\usage{
genes
}
\description{
A subset of the object \code{genes} from Bioconductor workflow RNAseq123.
}
\keyword{datasets}
