% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getAnnotations}
\alias{getAnnotations}
\alias{getAnnotations.oaStudy}
\alias{getAnnotations.SQLiteConnection}
\alias{getAnnotations.character}
\title{Get annotations from a study}
\usage{
getAnnotations(study, annotationID = NULL, ...)

\method{getAnnotations}{oaStudy}(study, annotationID = NULL, ...)

\method{getAnnotations}{SQLiteConnection}(study, annotationID = NULL, ...)

\method{getAnnotations}{character}(study, annotationID = NULL, libraries = NULL, ...)
}
\description{
Get annotations from a study
}
