% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getAssays}
\alias{getAssays}
\alias{getAssays.oaStudy}
\alias{getAssays.SQLiteConnection}
\alias{getAssays.character}
\title{Get assays from a study}
\usage{
getAssays(study, modelID = NULL, ...)

\method{getAssays}{oaStudy}(study, modelID = NULL, ...)

\method{getAssays}{SQLiteConnection}(study, modelID = NULL, ...)

\method{getAssays}{character}(study, modelID = NULL, libraries = NULL, ...)
}
\description{
Get assays from a study
}
