% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getBarcodeData}
\alias{getBarcodeData}
\title{Get data for barcode and violin plots}
\usage{
getBarcodeData(study, modelID, testID, annotationID, termID)
}
\description{
Get data for barcode and violin plots
}
