% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatabases.R
\name{getDatabases}
\alias{getDatabases}
\title{getDatabases}
\usage{
getDatabases(study, database)
}
\description{
getDatabases
}
\examples{
  getDatabases("***REMOVED***", "GOSLIM")[1:3, ]
  getDatabases("***REMOVED***", "msig_C6")[1:3, ]
  getDatabases("***REMOVED***", "PSP")[1:3, ]
}
