% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentNetwork.R
\name{getEnrichmentNetwork}
\alias{getEnrichmentNetwork}
\title{getEnrichmentNetwork}
\usage{
getEnrichmentNetwork(study, model, database)
}
\description{
getEnrichmentNetwork
}
\examples{
  getEnrichmentNetwork("***REMOVED***",
                       "No Pretreatment Timecourse Differential Phosphorylation",
                       "GO_BIOLOGICAL_PROCESS")
}
