% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentResults.R
\name{getEnrichmentResults}
\alias{getEnrichmentResults}
\title{getEnrichmentResults}
\usage{
getEnrichmentResults(model, database, study, type = "nominal")
}
\description{
getEnrichmentResults
}
\examples{
  getEnrichmentResults("No Pretreatment Timecourse Differential Phosphorylation",
                       "GO_BIOLOGICAL_PROCESS",
                       "***REMOVED***",
                       type = "nominal")
  getEnrichmentResults("No Pretreatment Timecourse Differential Phosphorylation",
                       "GO_BIOLOGICAL_PROCESS",
                       "***REMOVED***",
                       type = "adjusted")
}
