% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getEnrichments}
\alias{getEnrichments}
\alias{getEnrichments.oaStudy}
\alias{getEnrichments.SQLiteConnection}
\alias{getEnrichments.character}
\title{Get enrichments from a study}
\usage{
getEnrichments(study, modelID = NULL, annotationID = NULL, testID = NULL, ...)

\method{getEnrichments}{oaStudy}(study, modelID = NULL, annotationID = NULL, testID = NULL, ...)

\method{getEnrichments}{SQLiteConnection}(study, modelID = NULL, annotationID = NULL, testID = NULL, ...)

\method{getEnrichments}{character}(
  study,
  modelID = NULL,
  annotationID = NULL,
  testID = NULL,
  libraries = NULL,
  ...
)
}
\description{
Get enrichments from a study
}
