% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getEnrichmentsIntersection}
\alias{getEnrichmentsIntersection}
\title{getEnrichmentsIntersection}
\usage{
getEnrichmentsIntersection(
  study,
  modelID,
  annotationID,
  mustTests,
  notTests,
  sigValue,
  operator,
  type,
  ...
)
}
\description{
getEnrichmentsIntersection
}
