% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdowns.R
\name{getEnrichmentsModels}
\alias{getEnrichmentsModels}
\title{Get the models for the enrichments of an installed OmicNavigator study}
\usage{
getEnrichmentsModels(study, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Only accepts name of installed study
package.}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
A named list. The names are the modelIDs, and each list element is a
single character vector with the modelID description to be used as a
tooltip in the app. If no custom description was provided by the user, the
tooltip text is simply the modelID.
}
\description{
This is the API endpoint the app uses to populate the dropdown menu in the
Enrichment Analysis tab with the list of available models for the selected
study.
}
\details{
The models correspond to those used when adding the results with
\code{\link{addEnrichments}}. Any optional tooltips correspond to the
descriptions added with \code{\link{addModels}}.
}
\seealso{
\code{\link{getEnrichmentsStudies}},
\code{\link{getResultsModels}},
\code{\link{addEnrichments}},
\code{\link{addModels}}
}
