% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getEnrichmentsNetwork}
\alias{getEnrichmentsNetwork}
\alias{getEnrichmentsNetwork.SQLiteConnection}
\alias{getEnrichmentsNetwork.character}
\title{Get enrichments network from a study}
\usage{
getEnrichmentsNetwork(study, modelID, annotationID, ...)

\method{getEnrichmentsNetwork}{SQLiteConnection}(study, modelID, annotationID, ...)

\method{getEnrichmentsNetwork}{character}(study, modelID, annotationID, libraries = NULL, ...)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy}
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{annotationID}{Filter by annotation}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get enrichments network from a study
}
