% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getFavicons}
\alias{getFavicons}
\title{Get favicon URLs for table linkouts}
\usage{
getFavicons(linkouts)
}
\arguments{
\item{linkouts}{Character vector or (potentially nested) list of character
vectors containing the URLs for the table linkouts.}
}
\value{
The URLs to the favicons for each linkout. The output returned will
  always be the same class and structure as the input.
}
\description{
To enhance the display of the linkouts in the app's tables, it can fetch the
favicon URL for each website.
}
\examples{
  getFavicons("https://reactome.org/content/detail/")

}
\seealso{
\code{\link{getResultsLinkouts}},
         \code{\link{getEnrichmentsLinkouts}}
}
