% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getInferences}
\alias{getInferences}
\alias{getInferences.oaStudy}
\alias{getInferences.SQLiteConnection}
\alias{getInferences.character}
\title{Get inference results from a study}
\usage{
getInferences(study, model = NULL, contrast = NULL)

\method{getInferences}{oaStudy}(study, model = NULL, contrast = NULL)

\method{getInferences}{SQLiteConnection}(study, model = NULL, contrast = NULL)

\method{getInferences}{character}(study, model = NULL, contrast = NULL, libraries = NULL)
}
\description{
Get inference results from a study
}
