% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getLinkFeatures}
\alias{getLinkFeatures}
\title{Get the shared features in a network link}
\usage{
getLinkFeatures(study, annotationID, termID1, termID2)
}
\description{
Get the shared features in a network link
}
