% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getMetaFeaturesTable}
\alias{getMetaFeaturesTable}
\title{Get metaFeatures for a given feature}
\usage{
getMetaFeaturesTable(study, modelID, featureID)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{oaStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{featureID}{Filter by featureID}
}
\description{
Get metaFeatures for a given feature
}
