% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getModels}
\alias{getModels}
\alias{getModels.oaStudy}
\alias{getModels.SQLiteConnection}
\alias{getModels.character}
\title{Get models from a study}
\usage{
getModels(study, model = NULL)

\method{getModels}{oaStudy}(study, model = NULL)

\method{getModels}{SQLiteConnection}(study, model = NULL)

\method{getModels}{character}(study, model = NULL, libraries = NULL)
}
\description{
Get models from a study
}
