% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getOverlaps}
\alias{getOverlaps}
\title{Get overlaps from a study}
\usage{
getOverlaps(study, annotationID = NULL, libraries = NULL)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
or the name of an installed study package.}

\item{annotationID}{Filter by annotation}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get overlaps from a study
}
