% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getPackageVersion}
\alias{getPackageVersion}
\title{Get version of OmicNavigator package}
\usage{
getPackageVersion(libraries = NULL)
}
\arguments{
\item{libraries}{Directory path(s) to R package library(ies). Passed to
the argument \code{lib.loc} of \code{\link[utils]{packageVersion}}.}
}
\value{
Returns a one-element character vector with the version of the
currently installed OmicNavigator R package
}
\description{
This is a convenience function for the app. It is easier to always call the
OmicNavigator package functions via OpenCPU than to call the utils package
for this one endpoint.
}
\seealso{
\code{\link[utils]{packageVersion}}
}
