% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getPlots}
\alias{getPlots}
\alias{getPlots.oaStudy}
\alias{getPlots.SQLiteConnection}
\alias{getPlots.character}
\title{Get plots from a study}
\usage{
getPlots(study, modelID = NULL, ...)

\method{getPlots}{oaStudy}(study, modelID = NULL, ...)

\method{getPlots}{SQLiteConnection}(study, modelID = NULL, ...)

\method{getPlots}{character}(study, modelID = NULL, libraries = NULL, ...)
}
\description{
Get plots from a study
}
