% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getReportLink}
\alias{getReportLink}
\title{Get link to report}
\usage{
getReportLink(study, modelID)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}
}
\value{
Returns a one-element character vector with either a path to a report
file or a URL to a report web page. If no report is available for the
modelID, an empty character vector is returned.
}
\description{
Get link to report
}
