% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getReports}
\alias{getReports}
\title{Get reports from a study}
\usage{
getReports(study, modelID = NULL, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{oaStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get reports from a study
}
