% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getResultsIntersection}
\alias{getResultsIntersection}
\title{getResultsIntersection}
\usage{
getResultsIntersection(
  study,
  modelID,
  anchor,
  mustTests,
  notTests,
  sigValue,
  operator,
  column,
  ...
)
}
\description{
getResultsIntersection
}
