% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdowns.R
\name{getResultsStudies}
\alias{getResultsStudies}
\title{Get installed OmicNavigator studies that have results}
\usage{
getResultsStudies(libraries = NULL)
}
\arguments{
\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
Returns a character vector of the installed OmicNavigator study
packages
}
\description{
This is the API endpoint the app uses to populate the dropdown menu in the
Differential Analysis tab with the list of available studies with results
data.
}
\details{
Internally, \code{getResultsStudies} calls \code{\link{getInstalledStudies}}
with \code{hasElements = "results"}.
}
\seealso{
\code{\link{getInstalledStudies}},
\code{\link{getEnrichmentsStudies}}
}
