% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getResultsUpset}
\alias{getResultsUpset}
\title{getResultsUpset}
\usage{
getResultsUpset(study, modelID, sigValue, operator, column, legacy = FALSE)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{sigValue}{The numeric significance value to use as a cutoff for each
column}

\item{operator}{The comparison operators for each column, e.g. \code{"<"}}

\item{column}{The columns to apply the filters}

\item{legacy}{Use legacy code (for testing purposes only)}
}
\value{
Invisibly returns the output from \code{\link[UpSetR]{upset}}
}
\description{
getResultsUpset
}
