% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getSamples}
\alias{getSamples}
\alias{getSamples.oaStudy}
\alias{getSamples.SQLiteConnection}
\alias{getSamples.character}
\title{Get samples from a study}
\usage{
getSamples(study, modelID = NULL, ...)

\method{getSamples}{oaStudy}(study, modelID = NULL, ...)

\method{getSamples}{SQLiteConnection}(study, modelID = NULL, ...)

\method{getSamples}{character}(study, modelID = NULL, libraries = NULL, ...)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy}
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get samples from a study
}
