% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStudies.R
\name{getStudies}
\alias{getStudies}
\title{Return list of phosphoproteomics studies}
\usage{
getStudies(libraries = NULL)
}
\arguments{
\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
character vector of PhoshoApp study packages
}
\description{
Return list of phosphoproteomics studies
}
\examples{
 getStudies()
}
