% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getStudyMeta}
\alias{getStudyMeta}
\title{Get study metadata}
\usage{
getStudyMeta(study, libraries = NULL)
}
\arguments{
\item{study}{Name of an installed OmicNavigator study}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
Returns a list with the following components:

\item{description}{(character) Study description}
\item{version}{(character) Study version}
\item{maintainer}{(character) Study maintainer}
\item{maintainerEmail}{(character) Study maintainer email}
\item{studyMeta}{(list) Additional study metadata added via the argument
\code{studyMeta} of \code{\link{createStudy})}}
}
\description{
Get the study description, version, maintainer, maintainer email, and any
extra metadata added via the argument \code{studyMeta} of
\code{\link{createStudy}}.
}
\seealso{
\code{\link{createStudy}}
}
