% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTests}
\alias{getTests}
\alias{getTests.oaStudy}
\alias{getTests.SQLiteConnection}
\alias{getTests.character}
\title{Get tests from a study}
\usage{
getTests(study, modelID = NULL, ...)

\method{getTests}{oaStudy}(study, modelID = NULL, ...)

\method{getTests}{SQLiteConnection}(study, modelID = NULL, ...)

\method{getTests}{character}(study, modelID = NULL, libraries = NULL, ...)
}
\description{
Get tests from a study
}
