% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getUpsetCols}
\alias{getUpsetCols}
\title{getUpsetCols}
\usage{
getUpsetCols(study, modelID, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
Returns a character vector with the names of the common columns
}
\description{
Determine the common columns across all tests of a model that are available
for filtering with UpSet.
}
