% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group}
\alias{group}
\title{group from Bioconductor workflow RNAseq123}
\format{
A factor with 3 levels:
\describe{
  \item{Basal}{Basal cells}
  \item{LP}{Luminal progenitor cells}
  \item{ML}{Mature luminal cells}
}
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
group
}
\description{
A subset of the object \code{group} from Bioconductor workflow RNAseq123.
}
\examples{
  data("RNAseq123", package = "OmicAnalyzer")
  table(group)
  str(group)

}
\keyword{datasets}
