% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdowns.R
\name{hasTerms}
\alias{hasTerms}
\title{Does an annotationID have terms available?}
\usage{
hasTerms(study, annotationID, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Only accepts name of installed study
package.}

\item{annotationID}{The annotationID (selected in the app) to check for
available terms}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
\code{TRUE} if the annotationID has terms available; otherwise
\code{FALSE}.
}
\description{
Only populate the enrichments table with links to the barcode view if the
annotationID has terms avaialble.
}
\seealso{
\code{\link{getEnrichmentsAnnotations}},
\code{\link{addAnnotations}}
}
