% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importStudy}
\alias{importStudy}
\title{Import a study package}
\usage{
importStudy(study, libraries = NULL)
}
\arguments{
\item{study}{Named of an installed OmicNavigator study}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Create an onStudy object by importing an installed study package
}
