% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{installApp}
\alias{installApp}
\title{Install the OmicNavigator web app}
\usage{
installApp(version = NULL, overwrite = FALSE, lib.loc = NULL, ...)
}
\arguments{
\item{version}{Version of the web app to install, e.g. \code{"1.0.0"}}

\item{overwrite}{Should an existing installation of the app be overwritten?}

\item{lib.loc}{a character vector with path names of \R libraries.
    See \sQuote{Details} for the meaning of the default value of \code{NULL}.}

\item{...}{Passed to \code{\link[utils]{download.file}}. If the download
fails, you may need to adjust the download settings for your operating
system. For example, to download with \code{wget}, pass the argument
\code{method = "wget"}.}
}
\description{
In order to run the OmicNavigator web app on your local machine, the app
must be installed in the \code{www/} subdirectory of the R package. By default,
the app is downloaded automatically during installation of the R package. If
this download failed, or if you want to use a different version of the app,
you can manually download and install the app.
}
