% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lane}
\alias{lane}
\title{lane from Bioconductor workflow RNAseq123}
\format{
A data frame with 24 rows and 3 columns:
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
lane
}
\description{
A subset of the object \code{lane} from Bioconductor workflow RNAseq123.
}
\examples{
  data("RNAseq123", package = "OmicAnalyzer")
  head(lane)
  str(lane)

}
\keyword{datasets}
