% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcpm}
\alias{lcpm}
\title{lcpm from Bioconductor workflow RNAseq123}
\format{
A matrix with 24 rows and 9 columns
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
lcpm
}
\description{
A subset of the object \code{lcpm} from Bioconductor workflow RNAseq123.
}
\examples{
  data("RNAseq123", package = "OmicAnalyzer")
  head(lcpm)
  str(lcpm)

}
\keyword{datasets}
