% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{listStudies}
\alias{listStudies}
\title{List available studies and their metadata}
\usage{
listStudies(libraries = NULL)
}
\arguments{
\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
List available studies and their metadata
}
