% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mm.c2}
\alias{Mm.c2}
\title{Mm.c2 from Bioconductor workflow RNAseq123}
\format{
A list of 4 character vectors
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
Mm.c2
}
\description{
A subset of the object \code{Mm.c2} from Bioconductor workflow RNAseq123.
}
\examples{
 data("RNAseq123", package = "OmicAnalyzer")
 Mm.c2[[1]]
 str(Mm.c2)

}
\keyword{datasets}
