% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{OmicAnalyzer}
\alias{OmicAnalyzer}
\title{OmicAnalyzer}
\description{
Package options to control package-wide behavior are described below.
}
\details{
The default prefix for OmicAnalyzer study packages is "OAstudy". If you
would prefer to use a different prefix, you can change the package option
\code{OmicAnalyzer.prefix}. For example, to use the prefix "OmicAnalyzerStudy",
you could add the following line to your \code{.Rprofile} file.

\preformatted{
options(OmicAnalyzer.prefix = "OmicAnalyzerStudy")
}
}
