% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{OmicNavigator}
\alias{OmicNavigator}
\title{OmicNavigator}
\description{
Package options to control package-wide behavior are described below.
}
\details{
The default prefix for OmicNavigator study packages is "ONstudy". If you
would prefer to use a different prefix, you can change the package option
\code{OmicNavigator.prefix}. For example, to use the prefix "OmicNavigatorStudy",
you could add the following line to your \code{.Rprofile} file.

\preformatted{
options(OmicNavigator.prefix = "OmicNavigatorStudy")
}
}
