% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{removeStudy}
\alias{removeStudy}
\title{Remove an installed study R package}
\usage{
removeStudy(study, library = .libPaths()[1])
}
\arguments{
\item{study}{The name of the study or an \code{onStudy} object. Do \strong{not} include
the prefix of the installed package, e.g. \code{ONstudy}.}

\item{library}{Directory where the study package is installed. Defaults to first directory
returned by \code{\link{.libPaths}}.}
}
\value{
Invisibly returns the path of the removed study package
}
\description{
Remove an installed study R package
}
