% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplenames}
\alias{samplenames}
\title{samplenames from Bioconductor workflow RNAseq123}
\format{
A character vector containing the unique sample identifiers
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
samplenames
}
\description{
A subset of the object \code{samplenames} from Bioconductor workflow RNAseq123.
}
\examples{
 data("RNAseq123", package = "OmicAnalyzer")
 head(samplenames)
 str(samplenames)

}
\keyword{datasets}
