% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{shared-add}
\alias{shared-add}
\title{Shared parameters for add functions}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}
}
\description{
Shared parameters for add functions
}
\keyword{internal}
