% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{shared-data}
\alias{shared-data}
\title{Shared sections for data objects}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\description{
Shared sections for data objects
}
\references{
Law CW, Alhamdoosh M, Su S, Dong X, Tian L, Smyth GK, Ritchie ME.
  \href{https://f1000research.com/articles/5-1408/v3}{RNA-seq analysis is easy as 1-2-3 with limma, Glimma and edgeR [version 3; peer review: 3 approved].}
  F1000Research 2018, 5:1408

  Sheridan, J.M., Ritchie, M.E., Best, S.A. et al.
  \href{https://doi.org/10.1186/s12885-015-1187-z}{A pooled shRNA screen for regulators of primary mammary stem and progenitor cells identifies roles for \emph{Asap1} and \emph{Prox1}.}
  BMC Cancer 2015, 15:221
}
\keyword{internal}
