% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdowns.R
\name{shared-dropdowns}
\alias{shared-dropdowns}
\title{Shared parameters for dropdown functions}
\arguments{
\item{study}{An OmicNavigator study. Only accepts name of installed study
package.}

\item{modelID}{The modelID selected by the user in the app}
}
\value{
A named list. The names are the identifiers to be displayed in the
dropdown menu, and each list element is a single character vector with the
description to be used as a tooltip in the app. If no custom description
was provided by the user, the tooltip text is simply the identifier.
}
\description{
Shared parameters for dropdown functions
}
\keyword{internal}
