% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{shared-get}
\alias{shared-get}
\title{Shared parameters for get functions}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
a \code{\link[RSQLite]{SQLiteConnection-class}} to a study database, or the name
of an installed study package.}

\item{modelID}{Filter by modelID}

\item{testID}{Filter by testID}

\item{annotationID}{Filter by annotation}

\item{termID}{Filter by termID}

\item{featureID}{Filter by featureID}

\item{plotID}{Filter by plotID}

\item{...}{Currently unused but required for potential future arguments. See
the section
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Generic-functions-and-methods}{Generic
functions and methods} in
\href{https://cran.r-project.org/doc/manuals/R-exts.html}{Writing R
Extensions}.}
}
\description{
Shared parameters for get functions
}
\keyword{internal}
