% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{shared-upset}
\alias{shared-upset}
\title{Shared parameters for upset functions}
\arguments{
\item{anchor}{The primary testID to filter the results}

\item{mustTests}{The testIDs for which a featureID (or termID for enrichment)
must pass the filters}

\item{notTests}{The testIDs for which a featureID (or termID for enrichment)
must \bold{not} pass the filters. In other words, if a featureID passes the
filter for a testID specified in notTests, that featureID is removed from
the output}

\item{sigValue}{The numeric significance value to use as a cutoff for each
column}

\item{operator}{The comparison operators for each column, e.g. \code{"<"}}

\item{column}{The columns to apply the filters}

\item{type}{Type of p-value: (\code{"nominal"} or \code{"adjusted"})}

\item{tests}{Restrict UpSet plot to only include these tests}

\item{legacy}{Use legacy code (for testing purposes only)}
}
\description{
Shared parameters for upset functions
}
\keyword{internal}
