% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{startApp}
\alias{startApp}
\title{Start app on local machine}
\usage{
startApp(...)
}
\arguments{
\item{...}{extra parameters passed to \link[opencpu]{ocpu_start_server}}
}
\value{
No return value. This function is only called for the side effect of
  running a local instance of the app.
}
\description{
After you have installed at least one OmicNavigator study package with
\code{\link{installStudy}}, you can explore the results in the app. The
function \code{startApp} starts a local instance of the app running on your
current machine. It will automatically open the app in your default browser.
For the best experience, use Google Chrome. From the dropdown menu, you will
be able to select from any of the studies you have installed on your machine.
When you are finished, you can stop the web server by returning to the R
console and pressing the Esc key (Windows) or Ctrl-C (Linux, macOS).
}
\details{
Note that the app can't be run from within RStudio Server.

The app requires some additional R packages to run. If you receive an error
about a missing package, please install it with
\code{\link[utils]{install.packages}}. To ensure you have all the extra
packages installed, you can run the command below:

\preformatted{
install.packages(c("faviconPlease", "opencpu", "UpSetR"))
}
}
