% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.onStudy}
\alias{summary.onStudy}
\title{Summarize elements of OmicNavigator study}
\usage{
\method{summary}{onStudy}(object, elements = NULL, ...)
}
\arguments{
\item{object}{OmicNavigator study object (class \code{onStudy})}

\item{elements}{Subset the output to only include specific elements of the
study, e.g. \code{c("results", "enrichments")}}

\item{...}{Currently unused}
}
\value{
Invisibly returns the original \code{onStudy} object
}
\description{
Displays a tree-like summary of the elements that have been added to an
OmicNavigator study.
}
