% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateStudy}
\alias{validateStudy}
\title{Validate a study}
\usage{
validateStudy(study)
}
\arguments{
\item{study}{An OmicNavigator study object}
}
\value{
For a valid study object, the logical value \code{TRUE} is invisibly
returned. For an invalid study object, there is no return value because an
error is thrown.
}
\description{
Validate a study
}
