% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAnnotations}
\alias{addAnnotations}
\title{Add annotations}
\usage{
addAnnotations(study, annotations, overwrite = FALSE)
}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{annotations}{The annotations used for the enrichment analyses. The
input is a nested list. The top-level list contains one entry per
annotation database, e.g. reactome. The names correspond to the name of
each annotation database. Each of these elements should be list of that
contains more information about each annotation database. Specifically the
sublist should contain 1) \code{description}, a character vector that
describes the resource, 2) \code{featureID}, the name of the column in the
features table that was used for the enrichment analysis (if omitted, it is
assumed to be the main featureID used for the study), and 3) \code{terms},
a list of annotation terms. The names of \code{terms} sublist correspond to
the name of the annotation terms. Each of the annotation terms should be a
character vector of feature IDs.}

\item{overwrite}{Overwrite existing value}
}
\description{
Add annotations
}
