% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAssays}
\alias{addAssays}
\title{Add assays}
\usage{
addAssays(study, assays)
}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}

\item{assays}{The assays from the study. The input object is a list of data
frames (one per model). The row names should correspond to the featureIDs
(\code{\link{addFeatures}}). The column names should correspond to the
sampleIDs (\code{\link{addSamples}}). The data frame should only contain
numeric values. To share a data frame across multiple models, use the
modelID "default".}
}
\description{
Add assays
}
