% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput.R
\name{addAssays}
\alias{addAssays}
\title{Add assays}
\usage{
addAssays(study, assays, overwrite = FALSE)
}
\arguments{
\item{assays}{The assays from the study. The input is a named list. The names
of the list correspond to the model names. Each element in the list should
be a matrix of quantifications for the assays. The column names should be
the sample IDs and the rows should be the feature IDs.}
}
\description{
Add assays
}
