% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAssays}
\alias{addAssays}
\title{Add assays}
\usage{
addAssays(study, assays, reset = FALSE)
}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}

\item{assays}{The assays from the study. The input object is a list of data
frames (one per model). The row names should correspond to the featureIDs
(\code{\link{addFeatures}}). The column names should correspond to the
sampleIDs (\code{\link{addSamples}}). The data frame should only contain
numeric values. To share a data frame across multiple models, use the
modelID "default".}

\item{reset}{Reset the data prior to adding the new data (default:
\code{FALSE}). The default is to add to or modify any previously added data
(if it exists). Setting \code{reset = TRUE} enables you to remove existing
data you no longer want to include in the study.}
}
\value{
Returns the original \code{onStudy} object passed to the argument
\code{study}, but modified to include the newly added data
}
\description{
Add assays to the models of an OmicNavigator study.
}
\details{
If you want to add multiple transformations of your assays for a given
modelID, you can add another layer of nesting. Instead of passing a data
frame for that modelID, you can pass a named list of data frames. Each
transformation should have identical row and column names. If your data
doesn't fit this restriction, you should probably use multiple models
instead.
}
\seealso{
\code{\link{getAssays}}
}
