% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput.R
\name{addEnrichments}
\alias{addEnrichments}
\title{Add enrichment results}
\usage{
addEnrichments(study, enrichments, overwrite = FALSE)
}
\arguments{
\item{enrichments}{The enrichment results from each model. The input is a
nested named list. The names of the list correspond to the model names.
Each list element should be a list of the contrasts tested. The names
correspond to the contrast names. Each list element should be another list
of annotation databases. The names correspond to the annotation databases.
Each of these elements should be a data frame with enrichment results. Each
table must have a column named "key" that contains the annotation terms.}
}
\description{
Add enrichment results
}
