% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addFeatures}
\alias{addFeatures}
\title{Add feature metadata}
\usage{
addFeatures(study, features)
}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{features}{The metadata variables that describe the features in the
study. The input object is a list of data frames (one per model). The first
column of each data frame is used as the featureID, so it must contain
unique values. To share a data frame across multiple models, use the
modelID "default". All columns will be coerced to character strings.}
}
\description{
Add feature metadata
}
