% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addFeatures}
\alias{addFeatures}
\title{Add feature metadata}
\usage{
addFeatures(study, features, overwrite = FALSE)
}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{features}{The metadata variables that describe the features in the
study. The input object can be a single data frame or a list of data frames
(one per model). The first column is used as the feature ID, so it must
contain unique values.}

\item{overwrite}{Overwrite existing value}
}
\description{
Add feature metadata
}
