% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addMapping}
\alias{addMapping}
\title{Add mapping object}
\usage{
addMapping(study, mapping, reset = FALSE)
}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}

\item{mapping}{Feature IDs from models. The input object is a list of named
data frames. Column names indicate model names (modelID), and rows indicate
featureIDs per model. Features with same index position across columns are
treated as mapped across models. For each model, feature IDs must match
feature IDs available in the results object of the respective model.
The name of each list element should be set to either a modelID or "default".
E.g., if in a study there are models "transcriptomics" and "proteomics" and
the user wants to create a plot based on data from both, a mapping object
with column names "transcriptomics" and "proteomics" should be created, where
feature IDs of both models that relate to each other are located in the same
row across columns.}

\item{reset}{Reset the data prior to adding the new data (default:
\code{FALSE}). The default is to add to or modify any previously added data
(if it exists). Setting \code{reset = TRUE} enables you to remove existing
data you no longer want to include in the study.}
}
\value{
Returns the original \code{onStudy} object passed to the argument
  \code{study}, but modified to include the newly added data
}
\description{
Add mapping object
}
\seealso{
\code{\link{getPlottingData}}, \code{\link{plotStudy}}
}
