% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addMetaFeatures}
\alias{addMetaFeatures}
\title{Add meta-feature metadata}
\usage{
addMetaFeatures(study, metaFeatures)
}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{metaFeatures}{The metadata variables that describe the meta-features in
the study. The input object is a list of data frames (one per model). The
first column of each data frame is used as the featureID, so it must
contain the same IDs as the corresponding features data frame
(\code{\link{addFeatures}}). To share a data frame across multiple models,
use the modelID "default". All columns will be coerced to character
strings.}
}
\description{
The meta-features table is useful anytime there are metadata variables that
cannot be mapped 1:1 to your features. For example, a peptide may be
associated with multiple proteins.
}
