% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addModels}
\alias{addModels}
\title{Add models}
\usage{
addModels(study, models, overwrite = FALSE)
}
\arguments{
\item{models}{The models analyzed in the study. The input is a named
character vector. The names correspond to the names of the models. The
elements correspond to the descriptions of the models.}
}
\description{
Add models
}
