% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addResultsLinkouts}
\alias{addResultsLinkouts}
\title{Add linkouts to external resources in the results table}
\usage{
addResultsLinkouts(study, resultsLinkouts)
}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}

\item{resultsLinkouts}{The URL patterns that describe linkouts to external
resources (see Details below). The input object is a list of character
vectors (one per model). To share linkouts across multiple models, use the
modelID "default".}
}
\description{
You can provide additional information on the features in your study by
providing linkouts to external resources. These will be embedded directly in
the results table.
}
\details{
To construct a valid URL pattern to describe a linkout, you can include any
of the column names in the features table by enclosing them in \code{${}}. As
an example, if your features table included a column named \code{"ensembl"}
that contained the Ensembl Gene ID for each feature, you could create a
linkout to Ensembl using the following pattern:

\preformatted{"https://ensembl.org/Homo_sapiens/Gene/Summary?g=${ensembl}"}

As another example, if you had a column named \code{"entrez"} that contained
the Entrez Gene ID  for each feature, you could create a linkout to Entrez
using the following pattern:

\preformatted{"https://www.ncbi.nlm.nih.gov/gene/${entrez}"}
}
\examples{
  study <- createStudy("example")
  resultsLinkouts <- list(
    default = c("https://ensembl.org/Homo_sapiens/Gene/Summary?g=${ensembl}",
                "https://www.ncbi.nlm.nih.gov/gene/${entrez}")
  )
  study <- addResultsLinkouts(study, resultsLinkouts)

}
\seealso{
\code{\link{addFeatures}}
}
