% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addSamples}
\alias{addSamples}
\title{Add sample metadata}
\usage{
addSamples(study, samples, overwrite = FALSE)
}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{samples}{The metadata variables that describe the samples in the study.
The input object can be a single data frame or a list of data frames (one
per model). The first column is used as the sample ID, so it must contain
unique values.}

\item{overwrite}{Overwrite existing value}
}
\description{
Add sample metadata
}
