% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addSamples}
\alias{addSamples}
\title{Add sample metadata}
\usage{
addSamples(study, samples)
}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{samples}{The metadata variables that describe the samples in the study.
The input object is a named list of data frames (one per model). The first
column of each data frame is used as the sampleID, so it must contain
unique values. To share a data frame across multiple models, use the
modelID "default".}
}
\description{
Add sample metadata
}
