% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTests}
\alias{addTests}
\title{Add tests}
\usage{
addTests(study, tests)
}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{tests}{The tests from the study. The input object is a list of data
frames (one per model). The first column should contain the unique ID for
each test. The second column should contain a description of the test. Any
additional columns will be removed. The column names will be set to
"testID" and "desription". To share a data frame across multiple models,
use the model ID "default".}
}
\description{
Add tests
}
