% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTests}
\alias{addTests}
\title{Add tests}
\usage{
addTests(study, tests)
}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{tests}{The tests from the study. The input object is a list of lists.
Each element of the top-level list is a model. The names should be the
modelIDs. For each modelID, each element of the nested list is a test. The
names should be the testIDs. The value should be a single character string
describing the testID. To share tests across multiple models, use the
modelID "default".}
}
\description{
Add tests
}
\examples{
  study <- createStudy("example")
  tests <- list(
    default = list(
      test1 = "Name of first test",
      test2 = "Name of second test"
    )
  )
  study <- addTests(study, tests)

}
