% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTests}
\alias{addTests}
\title{Add tests}
\usage{
addTests(study, tests, overwrite = FALSE)
}
\arguments{
\item{tests}{The tests tested by the model(s). A named list.
The names correspond to the name of each test. The elements
correspond to the description of each test.}
}
\description{
Add tests
}
