% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basal.vs.ml}
\alias{basal.vs.ml}
\title{basal.vs.ml from Bioconductor workflow RNAseq123}
\format{
A data frame with 24 rows and 8 columns:
\describe{
  \item{ENTREZID}{Entrez ID of mouse gene}
  \item{SYMBOL}{Symbol of mouse gene}
  \item{TXCHROM}{Chromosome location of mouse gene}
  \item{logFC}{Log fold change}
  \item{AveExpr}{Average expression level of the gene across all samples}
  \item{t}{Moderated t-statistic}
  \item{P.Value}{p-value}
  \item{adj.P.Val}{Adjusted p-value}
}
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
basal.vs.ml
}
\description{
A subset of the object \code{basal.vs.ml} from Bioconductor workflow RNAseq123.
}
\examples{
  data("RNAseq123", package = "OmicNavigator")
  head(basal.vs.ml)
  str(basal.vs.ml)

}
\references{
Law CW, Alhamdoosh M, Su S, Dong X, Tian L, Smyth GK, Ritchie ME.
  \href{https://f1000research.com/articles/5-1408/v3}{RNA-seq analysis is easy as 1-2-3 with limma, Glimma and edgeR [version 3; peer review: 3 approved].}
  F1000Research 2018, 5:1408

  Sheridan, J.M., Ritchie, M.E., Best, S.A. et al.
  \href{https://doi.org/10.1186/s12885-015-1187-z}{A pooled shRNA screen for regulators of primary mammary stem and progenitor cells identifies roles for \emph{Asap1} and \emph{Prox1}.}
  BMC Cancer 2015, 15:221
}
\keyword{datasets}
