% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cam.BasalvsML}
\alias{cam.BasalvsML}
\title{cam.BasalvsML from Bioconductor workflow RNAseq123}
\format{
A data frame with 4 rows and 4 columns:
\describe{
\item{NGenes}{Number of genes in each term}
\item{Direction}{Direction of the enrichment}
\item{PValue}{Nominal p-value}
\item{FDR}{Multiple-testing adjusted p-value}
}
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
cam.BasalvsML
}
\description{
A subset of the object \code{cam.BasalvsML} from Bioconductor workflow RNAseq123.
}
\examples{
  head(cam.BasalvsML)
  str(cam.BasalvsML)

}
\references{
Law CW, Alhamdoosh M, Su S, Dong X, Tian L, Smyth GK, Ritchie ME.
\href{https://f1000research.com/articles/5-1408/v3}{RNA-seq analysis is easy as 1-2-3 with limma, Glimma and edgeR \link{version 3; peer review: 3 approved}.}
F1000Research 2018, 5:1408
\doi{10.12688/f1000research.9005.3}

Sheridan, J.M., Ritchie, M.E., Best, S.A. et al.
\href{https://bmccancer.biomedcentral.com/articles/10.1186/s12885-015-1187-z}{A pooled shRNA screen for regulators of primary mammary stem and progenitor cells identifies roles for \emph{Asap1} and \emph{Prox1}.}
BMC Cancer 2015, 15:221
\doi{10.1186/s12885-015-1187-z}
}
\keyword{datasets}
