% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{createStudy}
\alias{createStudy}
\title{Create a study}
\usage{
createStudy(
  name,
  description = name,
  samples = list(),
  features = list(),
  models = list(),
  assays = list(),
  tests = list(),
  annotations = list(),
  results = list(),
  enrichments = list(),
  metaFeatures = list(),
  plots = list(),
  barcodes = list(),
  version = NULL
)
}
\arguments{
\item{name}{Name of the study}

\item{description}{Description of the study}

\item{samples}{The metadata variables that describe the samples in the study.
The input object is a named list of data frames (one per model). The first
column of each data frame is used as the sample ID, so it must contain
unique values. To share a data frame across multiple models, use the model
ID "default".}

\item{features}{The metadata variables that describe the features in the
study. The input object is a list of data frames (one per model). The first
column of each data frame is used as the feature ID, so it must contain
unique values. To share a data frame across multiple models, use the model
ID "default". All columns will be coerced to character strings.}

\item{models}{The models analyzed in the study. The input is a named list.
The names correspond to the names of the models. The elements correspond to
the descriptions of the models.}

\item{assays}{The assays from the study. The input object is a list of data
frames (one per model). The row names should correspond to the feature IDs
(\code{\link{addFeatures}}). The column names should corresond to the
sample IDs (\code{\link{addSamples}}). The data frame should only contain
numeric values. To share a data frame across multiple models, use the model
ID "default".}

\item{tests}{The tests from the study. The input object is a list of data
frames (one per model). The first column should contain the unique ID for
each test. The second column should contain a description of the test. Any
additional columns will be removed. The column names will be set to
"testID" and "desription". To share a data frame across multiple models,
use the model ID "default".}

\item{annotations}{The annotations used for the enrichment analyses. The
input is a nested list. The top-level list contains one entry per
annotation database, e.g. reactome. The names correspond to the name of
each annotation database. Each of these elements should be list of that
contains more information about each annotation database. Specifically the
sublist should contain 1) \code{description}, a character vector that
describes the resource, 2) \code{featureID}, the name of the column in the
features table that was used for the enrichment analysis, and 3) \code{terms},
a list of annotation terms. The names of \code{terms} sublist correspond to
the name of the annotation terms. Each of the annotation terms should be a
character vector of feature IDs.}

\item{results}{The inference results from each model. The input is a
nested named list. The names of the list correspond to the model names.
Each element in the list should be a list of data frames with inference
results, one for each test. In each data frame, the feature ID must be in
the first column, and all other columns must be numeric.}

\item{enrichments}{The enrichment results from each model. The input is a
nested named list. The names of the list correspond to the model names.
Each list element should be a list of the annotation databases tested
(\code{\link{addAnnotations}}). The names of the list correspond to the
annotation databases. Each list element should be another list of tests
(\code{\link{addTests}}). The names correspond to the tests performed. Each
of these elements should be a data frame with enrichment results. Each
table must contain the following columns: "termID", "description",
"nominal" (the nominal statistics), and "adjusted" (the statistics after
adjusting for multiple testing). Any additional columns are ignored.}

\item{metaFeatures}{The metadata variables that describe the meta-features in
the study. The input object is a list of data frames (one per model). The
first column of each data frame is used as the feature ID, so it must
contain the same IDs as the corresponding features data frame
(\code{\link{addFeatures}}). To share a data frame across multiple models,
use the model ID "default".}

\item{plots}{Custom plotting functions for the study. The input object is a
named list of lists (one per model). The names of the sublists should
correspond to functions defined in the current R session. Each sublist must
define a \code{displayName} to control how the plot will be named in the
app. Optionally, if the plottting function requires external packages,
these can be defined in the element \code{packages}. To share plots across
multiple models, use the model ID "default".}

\item{barcodes}{The metadata variables that describe the barcode plot.
The input object is a list of lists (one per model). Each sublist must
contain the element \code{statistic}, which is the column name in the
results table to use to construct the barcode plot. Each sublist may
additionally contain any of the following optional elements:
1) \code{absolute} - Should the statistic be converted to its absolute
value (default is \code{TRUE}).
2) \code{logFoldChange} - The column name in the results table that contains
the log fold change values.
3) \code{labelStat} - The x-axis label to describe the statistic.
4) \code{labelLow} - The left-side label to describe low values of the statistic.
5) \code{labelHigh} - The right-side label to describe high values of the statistic.
6) \code{featureDisplay} - The feature variable to use to label the barcode plot
on hover.
To share metadata across multiple models, use the model ID "default".}

\item{version}{(Optional) Include a version number to track the updates to
your study package. If you export the study to a package, the version is
used as the package version.}
}
\description{
Create a study
}
\examples{

study <- createStudy(name = "ABC",
                     description = "An analysis of ABC")

}
\seealso{
\code{\link{addSamples}},
  \code{\link{addFeatures}},
  \code{\link{addModels}},
  \code{\link{addAssays}},
  \code{\link{addTests}},
  \code{\link{addAnnotations}},
  \code{\link{addResults}},
  \code{\link{addEnrichments}},
  \code{\link{addMetaFeatures}},
  \code{\link{addPlots}},
  \code{\link{addBarcodes}},
  \code{\link{exportStudy}},
  \code{\link{installStudy}}
}
