% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput.R
\name{createStudy}
\alias{createStudy}
\title{Create a study}
\usage{
createStudy(
  name,
  description = name,
  samples = NULL,
  features = NULL,
  models = NULL,
  assays = NULL,
  contrasts = NULL,
  annotations = NULL,
  inferences = NULL,
  enrichments = NULL,
  featureID = "featureID",
  sampleID = "sampleID"
)
}
\arguments{
\item{name}{Name of the study}

\item{description}{Description of the study}

\item{samples}{A table of metadata variables that describe the samples in the
study. The table must contain the unique sampleID used for the study. Also,
the object must inherit from the class data.frame.}

\item{features}{A table of metadata variables that describe the features in the
study. The table must contain the unique featureID used for the study. Also,
the object must inherit from the class data.frame.}

\item{models}{The models analyzed in the study. The input is a named
character vector. The names correspond to the names of the models. The
elements correspond to the descriptions of the models.}

\item{assays}{The assays from the study. The input is a named list. The names
of the list correspond to the model names. Each element in the list should
be a matrix of quantifications for the assays. The column names should be
the sample IDs and the rows should be the feature IDs.}

\item{contrasts}{The contrasts tested by the model(s). A named character
vector. The names correspond to the name of each contrast. The elements
correspond to the description of each contrast.}

\item{annotations}{The annotations used for the enrichment analyses. The input
is a nested list. The top-level list contains one entry per annotation
database, e.g. reactome. The names correspond to the name of each annotation
database. Each of these elements should be list of annotation terms. The
names of this sublist correspond to the name of the annotation term. Each of
the annotation terms should be a character vector of feature IDs.}

\item{featureID}{The column name that contains the unique identifiers for the
features measured in the study}

\item{sampleID}{The column name that contains the unique identifiers for the
samples measured in the study}
}
\description{
Create a study
}
\examples{

study <- createStudy(name = "ABC",
                     description = "An analysis of ABC")

}
