% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{createStudy}
\alias{createStudy}
\title{Create a study}
\usage{
createStudy(
  name,
  description = name,
  samples = NULL,
  features = NULL,
  models = NULL,
  assays = NULL,
  contrasts = NULL,
  annotations = NULL,
  inferences = NULL,
  enrichments = NULL,
  metaFeatures = NULL,
  plots = NULL,
  featureID = "featureID",
  sampleID = "sampleID"
)
}
\arguments{
\item{name}{Name of the study}

\item{description}{Description of the study}

\item{samples}{A table of metadata variables that describe the samples in the
study. The table must contain the unique sampleID used for the study. Also,
the object must inherit from the class data.frame.}

\item{features}{A table of metadata variables that describe the features in the
study. The table must contain the unique featureID used for the study. Also,
the object must inherit from the class data.frame.}

\item{models}{The models analyzed in the study. The input is a named
character vector. The names correspond to the names of the models. The
elements correspond to the descriptions of the models.}

\item{assays}{The assays from the study. The input is a named list. The names
of the list correspond to the model names. Each element in the list should
be a matrix of quantifications for the assays. The column names should be
the sample IDs and the rows should be the feature IDs.}

\item{contrasts}{The contrasts tested by the model(s). A named character
vector. The names correspond to the name of each contrast. The elements
correspond to the description of each contrast.}

\item{annotations}{The annotations used for the enrichment analyses. The
input is a nested list. The top-level list contains one entry per
annotation database, e.g. reactome. The names correspond to the name of
each annotation database. Each of these elements should be list of that
contains more information about each annotation database. Specifically the
sublist should contain 1) \code{description}, a character vector that
describes the resource, 2) \code{featureID}, the name of the column in the
features table that was used for the enrichment analysis (if omitted, it is
assumed to be the main featureID used for the study), and 3) \code{terms},
a list of annotation terms. The names of \code{terms} sublist correspond to
the name of the annotation terms. Each of the annotation terms should be a
character vector of feature IDs.}

\item{inferences}{The inference results from each model. The input is a
nested named list. The names of the list correspond to the model names.
Each element in the list should be a list of data frames with inference
results, one for each contrast. The featureID column needs to be included
in each table.}

\item{enrichments}{The enrichment results from each model. The input is a
nested named list. The names of the list correspond to the model names.
Each list element should be a list of the contrasts tested. The names
correspond to the contrast names. Each list element should be another list
of annotation databases. The names correspond to the annotation databases.
Each of these elements should be a data frame with enrichment results. Each
table must have a column named "termID" that contains the annotation terms.}

\item{metaFeatures}{A table of metadata variables that describe the
meta-features in the study. This is useful anytime there are metadata
variables that cannot be mapped 1:1 to your features. For example, a
peptide may be associated with multiple proteins. The table must contain
the unique featureID used for the study. Also, the object must inherit from
the class data.frame.}

\item{plots}{Custom plotting functions. The input is a nested list. Each
element of the list defines a custom plotting function via a list with
multiple options. The only required option is \code{definition}, which
contains the function definition. You can optionally include
\code{displayName} to control how the plot will be named in the app.
Lastly, if the plottting function requires external packages, these can be
defined in the argument \code{packages}.}

\item{featureID}{The column name that contains the unique identifiers for the
features measured in the study}

\item{sampleID}{The column name that contains the unique identifiers for the
samples measured in the study}
}
\description{
Create a study
}
\examples{

study <- createStudy(name = "ABC",
                     description = "An analysis of ABC")

}
