% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportStudy}
\alias{exportStudy}
\title{Export a study}
\usage{
exportStudy(
  study,
  type = c("rds", "text", "package"),
  path = NULL,
  requireValid = TRUE
)
}
\arguments{
\item{study}{An OmicAnalyzer study}

\item{type}{Export to a RDS file ("rds"), text files ("text"), or an
R package ("package")}

\item{path}{Optional file path to save the object}

\item{requireValid}{Require that study is valid before exporting}
}
\description{
Export a study
}
\seealso{
\code{\link{validateStudy}}
}
