% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getAnnotations}
\alias{getAnnotations}
\title{Get annotations from a study}
\usage{
getAnnotations(study, annotationID = NULL, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{annotationID}{Filter by annotation}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get annotations from a study
}
