% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getAssays}
\alias{getAssays}
\title{Get assays from a study}
\usage{
getAssays(study, modelID = NULL, quiet = FALSE, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{quiet}{Suppress messages (default: \code{FALSE})}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\value{
The object returned depends on the data available and any filters
  (e.g. the argument  \code{modelID}):

If no filters are specified, then the object returned is a nested list,
similar to the original input object.

If one or more filters are applied, then only a subset of the original nested
list is returned. Technically, each filter applied is used to subset the
original nested list using \code{\link[base]{[[}}.

If no data is available, an empty list is returned (\code{list()}).
}
\description{
Get assays from a study
}
