% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getBarcodeData}
\alias{getBarcodeData}
\title{Get data for barcode and violin plots}
\usage{
getBarcodeData(study, modelID, testID, annotationID, termID)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{testID}{Filter by testID}

\item{annotationID}{Filter by annotation}

\item{termID}{Filter by termID}
}
\description{
Get data for barcode and violin plots
}
