% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getBarcodes}
\alias{getBarcodes}
\alias{getBarcodes.oaStudy}
\alias{getBarcodes.SQLiteConnection}
\alias{getBarcodes.character}
\title{Get barcodes from a study}
\usage{
getBarcodes(study, modelID = NULL, ...)

\method{getBarcodes}{oaStudy}(study, modelID = NULL, ...)

\method{getBarcodes}{SQLiteConnection}(study, modelID = NULL, ...)

\method{getBarcodes}{character}(study, modelID = NULL, libraries = NULL, ...)
}
\description{
Get barcodes from a study
}
