% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getEnrichments}
\alias{getEnrichments}
\title{Get enrichments from a study}
\usage{
getEnrichments(
  study,
  modelID = NULL,
  annotationID = NULL,
  testID = NULL,
  libraries = NULL
)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{annotationID}{Filter by annotation}

\item{testID}{Filter by testID}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get enrichments from a study
}
