% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getEnrichmentsIntersection}
\alias{getEnrichmentsIntersection}
\title{getEnrichmentsIntersection}
\usage{
getEnrichmentsIntersection(
  study,
  modelID,
  annotationID,
  mustTests,
  notTests,
  sigValue,
  operator,
  type,
  ...
)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{annotationID}{Filter by annotation}

\item{mustTests}{The tests whose significant values must be included. (The intersection)}

\item{notTests}{The tests whose significant values will be removed. (The difference)}

\item{sigValue}{The significance levels for each column.}

\item{operator}{The operators for each column.}

\item{type}{Type of p-value (\code{"nominal"} or \code{"adjusted"})}

\item{...}{Currently unused but required for potential future arguments. See
the section
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Generic-functions-and-methods}{Generic
functions and methods} in
\href{https://cran.r-project.org/doc/manuals/R-exts.html}{Writing R
Extensions}.}
}
\description{
getEnrichmentsIntersection
}
