% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getEnrichmentsNetwork}
\alias{getEnrichmentsNetwork}
\title{Get enrichments network from a study}
\usage{
getEnrichmentsNetwork(study, modelID, annotationID, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{annotationID}{Filter by annotationID}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\value{
Returns a list with the following components:

\item{tests}{(character) Vector of testIDs}
\item{nodes}{(data frame) The description of each annotation term (i.e.
node). The nominal and adjusted p-values are in list-columns.}
\item{links}{(list) The statistics for each pairwise overlap between the
annotation terms (i.e. nodes)}
}
\description{
Get enrichments network from a study
}
