% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getEnrichmentsNetwork}
\alias{getEnrichmentsNetwork}
\alias{getEnrichmentsNetwork.SQLiteConnection}
\alias{getEnrichmentsNetwork.character}
\title{Get enrichments network from a study}
\usage{
getEnrichmentsNetwork(study, modelID, annotationID, ...)

\method{getEnrichmentsNetwork}{SQLiteConnection}(study, modelID, annotationID, ...)

\method{getEnrichmentsNetwork}{character}(study, modelID, annotationID, libraries = NULL, ...)
}
\description{
Get enrichments network from a study
}
