% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getEnrichmentsTable}
\alias{getEnrichmentsTable}
\alias{getEnrichmentsTable.oaStudy}
\alias{getEnrichmentsTable.SQLiteConnection}
\alias{getEnrichmentsTable.character}
\title{Get enrichments table from a study}
\usage{
getEnrichmentsTable(study, modelID, annotationID, ...)

\method{getEnrichmentsTable}{oaStudy}(study, modelID, annotationID, ...)

\method{getEnrichmentsTable}{SQLiteConnection}(study, modelID, annotationID, ...)

\method{getEnrichmentsTable}{character}(study, modelID, annotationID, libraries = NULL, ...)
}
\description{
Get enrichments table from a study
}
