% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getEnrichmentsTable}
\alias{getEnrichmentsTable}
\alias{getEnrichmentsTable.oaStudy}
\alias{getEnrichmentsTable.SQLiteConnection}
\alias{getEnrichmentsTable.character}
\title{Get enrichments table from a study}
\usage{
getEnrichmentsTable(study, modelID, annotationID, type = "nominal", ...)

\method{getEnrichmentsTable}{oaStudy}(study, modelID, annotationID, type = "nominal", ...)

\method{getEnrichmentsTable}{SQLiteConnection}(study, modelID, annotationID, type = "nominal", ...)

\method{getEnrichmentsTable}{character}(
  study,
  modelID,
  annotationID,
  type = "nominal",
  libraries = NULL,
  ...
)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy}
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{annotationID}{Filter by annotation}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get enrichments table from a study
}
