% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getEnrichmentsUpset}
\alias{getEnrichmentsUpset}
\title{getEnrichmentsUpset}
\usage{
getEnrichmentsUpset(
  study,
  modelID,
  annotationID,
  sigValue,
  operator,
  type,
  ...
)
}
\description{
getEnrichmentsUpset
}
