% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getEnrichmentsUpset}
\alias{getEnrichmentsUpset}
\title{getEnrichmentsUpset}
\usage{
getEnrichmentsUpset(
  study,
  modelID,
  annotationID,
  sigValue,
  operator,
  type,
  tests = NULL
)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{annotationID}{Filter by annotationID}

\item{sigValue}{The numeric significance value to use as a cutoff for each
column}

\item{operator}{The comparison operators for each column, e.g. \code{"<"}}

\item{type}{Type of p-value: (\code{"nominal"} or \code{"adjusted"})}

\item{tests}{Restrict UpSet plot to only include these tests}
}
\value{
No return value. This function is called for the side effect of
creating an UpSet plot.
}
\description{
getEnrichmentsUpset
}
