% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getEnrichmentsUpset}
\alias{getEnrichmentsUpset}
\title{getEnrichmentsUpset}
\usage{
getEnrichmentsUpset(
  study,
  modelID,
  annotationID,
  sigValue,
  operator,
  type,
  tests = NULL
)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{annotationID}{Filter by annotation}

\item{sigValue}{The significance levels for each column.}

\item{operator}{The operators for each column.}

\item{type}{Type of p-value (\code{"nominal"} or \code{"adjusted"})}

\item{tests}{Restrict UpSet plot to these tests}
}
\description{
getEnrichmentsUpset
}
