% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures.oaStudy}
\alias{getFeatures.SQLiteConnection}
\alias{getFeatures.character}
\title{Get features from a study}
\usage{
getFeatures(study, modelID = NULL, ...)

\method{getFeatures}{oaStudy}(study, modelID = NULL, ...)

\method{getFeatures}{SQLiteConnection}(study, modelID = NULL, ...)

\method{getFeatures}{character}(study, modelID = NULL, libraries = NULL, ...)
}
\description{
Get features from a study
}
