% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getFeatures}
\alias{getFeatures}
\title{Get features from a study}
\usage{
getFeatures(study, modelID = NULL, quiet = FALSE, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{quiet}{Suppress messages (default: \code{FALSE})}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
A data frame (if \code{modelID} is specified) or a list of data
frames. All the columns will be character strings, even if the values
appear numeric.
}
\description{
Get features from a study
}
