% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getFeatures}
\alias{getFeatures}
\title{Get features from a study}
\usage{
getFeatures(study, modelID = NULL, libraries = NULL)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\value{
A data frame (if \code{modelID} is specified) or a list of data
  frames. All the columns will be character strings, even if the values
  appear numeric.
}
\description{
Get features from a study
}
