% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getInstalledStudies}
\alias{getInstalledStudies}
\title{Get installed OmicNavigator studies}
\usage{
getInstalledStudies(hasElements = NULL, libraries = NULL)
}
\arguments{
\item{hasElements}{Character vector of elements that must be present in the
study packages. Valid elements are 'metaFeatures', 'results',
'enrichments', 'reports', 'plots', 'assays', 'samples', 'features',
'resultsLinkouts', and 'metaAssays'. If \code{NULL} (default), then all
installed OmicNavigator studies are returned, regardless of their contents.}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
Returns a character vector of the installed OmicNavigator study
packages
}
\description{
Get installed OmicNavigator studies
}
