% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getLinkFeatures}
\alias{getLinkFeatures}
\title{Get the shared features in a network link}
\usage{
getLinkFeatures(study, annotationID, termID1, termID2, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Only accepts name of installed study
package.}

\item{annotationID}{Filter by annotationID}

\item{termID1, termID2}{Linked terms to find overlapping features}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\value{
Returns a character vector with the features included in both termIDs
(i.e. the intersection)
}
\description{
Get the shared features in a network link
}
\seealso{
\code{\link{getNodeFeatures}}
}
