% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getLinkFeatures}
\alias{getLinkFeatures}
\title{Get the shared features in a network link}
\usage{
getLinkFeatures(study, annotationID, termID1, termID2)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
a \code{\link[RSQLite]{SQLiteConnection-class}} to a study database, or the name
of an installed study package.}

\item{annotationID}{Filter by annotation}

\item{termID1, termID2}{Linked terms to find overlapping features}
}
\description{
Get the shared features in a network link
}
