% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinkFeatures.R
\name{getLinkFeatures}
\alias{getLinkFeatures}
\title{getLinkFeatures}
\usage{
getLinkFeatures(study, database, node1, node2)
}
\description{
getLinkFeatures
}
\examples{
  getLinkFeatures("***REMOVED***", "GO_BIOLOGICAL_PROCESS", "GO:0000027", "GO:0000184")
  getLinkFeatures("***REMOVED***", "GO_BIOLOGICAL_PROCESS", "GO:0000027", "GO:0006364")
}
