% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getModels}
\alias{getModels}
\alias{getModels.oaStudy}
\alias{getModels.SQLiteConnection}
\alias{getModels.character}
\title{Get models from a study}
\usage{
getModels(study, modelID = NULL, ...)

\method{getModels}{oaStudy}(study, modelID = NULL, ...)

\method{getModels}{SQLiteConnection}(study, modelID = NULL, ...)

\method{getModels}{character}(study, modelID = NULL, libraries = NULL, ...)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
a \code{\link[RSQLite]{SQLiteConnection-class}} to a study database, or the name
of an installed study package.}

\item{modelID}{Filter by modelID}

\item{...}{Currently unused but required for potential future arguments. See
the section
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Generic-functions-and-methods}{Generic
functions and methods} in
\href{https://cran.r-project.org/doc/manuals/R-exts.html}{Writing R
Extensions}.}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get models from a study
}
