% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getModels}
\alias{getModels}
\alias{getModels.oaStudy}
\alias{getModels.SQLiteConnection}
\alias{getModels.character}
\title{Get models from a study}
\usage{
getModels(study, modelID = NULL, ...)

\method{getModels}{oaStudy}(study, modelID = NULL, ...)

\method{getModels}{SQLiteConnection}(study, modelID = NULL, ...)

\method{getModels}{character}(study, modelID = NULL, libraries = NULL, ...)
}
\description{
Get models from a study
}
