% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getNodeFeatures}
\alias{getNodeFeatures}
\title{Get the features in a network node}
\usage{
getNodeFeatures(study, annotationID, termID, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{annotationID}{Filter by annotationID}

\item{termID}{Filter by termID}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
Returns a character vector with the features in the termID
}
\description{
Get the features in a network node
}
\seealso{
\code{\link{getLinkFeatures}}
}
