% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeFeatures.R
\name{getNodeFeatures}
\alias{getNodeFeatures}
\title{getNodeFeatures}
\usage{
getNodeFeatures(study, database, node)
}
\description{
getNodeFeatures
}
\examples{
  getNodeFeatures("***REMOVED***", "GO_BIOLOGICAL_PROCESS", "GO:0000002")
  getNodeFeatures("***REMOVED***", "GOSLIM", "GO:0140014")
}
