% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getNodeFeatures}
\alias{getNodeFeatures}
\title{Get the features in a network node}
\usage{
getNodeFeatures(study, annotationID, termID, libraries = NULL)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
a \code{\link[RSQLite]{SQLiteConnection-class}} to a study database, or the name
of an installed study package.}

\item{annotationID}{Filter by annotation}

\item{termID}{Filter by termID}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get the features in a network node
}
