% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getReportLink}
\alias{getReportLink}
\title{Get link to report}
\usage{
getReportLink(study, modelID)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{oaStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}
}
\description{
Get link to report
}
