% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getResults}
\alias{getResults}
\alias{getResults.oaStudy}
\alias{getResults.SQLiteConnection}
\alias{getResults.character}
\title{Get result results from a study}
\usage{
getResults(study, modelID = NULL, contrastID = NULL, ...)

\method{getResults}{oaStudy}(study, modelID = NULL, contrastID = NULL, ...)

\method{getResults}{SQLiteConnection}(study, modelID = NULL, contrastID = NULL, ...)

\method{getResults}{character}(study, modelID = NULL, contrastID = NULL, libraries = NULL, ...)
}
\description{
Get result results from a study
}
