% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getResultsTable}
\alias{getResultsTable}
\alias{getResultsTable.oaStudy}
\alias{getResultsTable.SQLiteConnection}
\alias{getResultsTable.character}
\title{Get results table from a study}
\usage{
getResultsTable(study, modelID, testID, ...)

\method{getResultsTable}{oaStudy}(study, modelID, testID, ...)

\method{getResultsTable}{SQLiteConnection}(study, modelID, testID, ...)

\method{getResultsTable}{character}(study, modelID, testID, libraries = NULL, ...)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy}
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{testID}{Filter by testID}

\item{...}{Currently unused but required for potential future arguments. See
the section
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Generic-functions-and-methods}{Generic
functions and methods} in
\href{https://cran.r-project.org/doc/manuals/R-exts.html}{Writing R
Extensions}.}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\description{
Get results table from a study
}
