% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getResultsTable}
\alias{getResultsTable}
\alias{getResultsTable.oaStudy}
\alias{getResultsTable.SQLiteConnection}
\alias{getResultsTable.character}
\title{Get results table from a study}
\usage{
getResultsTable(study, modelID, testID, ...)

\method{getResultsTable}{oaStudy}(study, modelID, testID, ...)

\method{getResultsTable}{SQLiteConnection}(study, modelID, testID, ...)

\method{getResultsTable}{character}(study, modelID, testID, libraries = NULL, ...)
}
\description{
Get results table from a study
}
