% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getResultsTable}
\alias{getResultsTable}
\title{Get results table from a study}
\usage{
getResultsTable(study, modelID, testID, libraries = NULL)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{testID}{Filter by testID}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\value{
A data frame which includes the columns from the features table
  followed by the columns from the results table. All the columns from the
  features table will be character strings, even if the values appear
  numeric.
}
\description{
Get results table from a study
}
