% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getResultsUpset}
\alias{getResultsUpset}
\title{getResultsUpset}
\usage{
getResultsUpset(study, modelID, sigValue, operator, column)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{sigValue}{The significance levels for each column.}

\item{operator}{The operators for each column.}

\item{column}{The columns to be thresheld.}
}
\description{
getResultsUpset
}
