% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getResultsUpset}
\alias{getResultsUpset}
\title{getResultsUpset}
\usage{
getResultsUpset(study, modelID, sigValue, operator, column, ...)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
a \code{\link[RSQLite]{SQLiteConnection-class}} to a study database, or the name
of an installed study package.}

\item{modelID}{Filter by modelID}

\item{sigValue}{The significance levels for each column.}

\item{operator}{The operators for each column.}

\item{column}{The columns to be thresheld.}

\item{...}{Currently unused but required for potential future arguments. See
the section
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Generic-functions-and-methods}{Generic
functions and methods} in
\href{https://cran.r-project.org/doc/manuals/R-exts.html}{Writing R
Extensions}.}
}
\description{
getResultsUpset
}
