% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getSamples}
\alias{getSamples}
\alias{getSamples.oaStudy}
\alias{getSamples.SQLiteConnection}
\alias{getSamples.character}
\title{Get samples from a study}
\usage{
getSamples(study, modelID = NULL, ...)

\method{getSamples}{oaStudy}(study, modelID = NULL, ...)

\method{getSamples}{SQLiteConnection}(study, modelID = NULL, ...)

\method{getSamples}{character}(study, modelID = NULL, libraries = NULL, ...)
}
\description{
Get samples from a study
}
