% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getStudies}
\alias{getStudies}
\title{Get installed OmicAnalyzer studies}
\usage{
getStudies(libraries = NULL)
}
\arguments{
\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
character vector of OmicAnalyzer study packages
}
\description{
Get installed OmicAnalyzer studies
}
\examples{
 getStudies()

}
