% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getStudyMeta}
\alias{getStudyMeta}
\title{Get study metadata from the study name}
\usage{
getStudyMeta(name)
}
\arguments{
\item{name}{Name of the study (without the OmicNavigator prefix)}
}
\value{
A list containing the study metadata
}
\description{
Get study metadata from the study name
}
