% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTests}
\alias{getTests}
\title{Get tests from a study}
\usage{
getTests(study, modelID = NULL, testID = NULL, quiet = FALSE, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{testID}{Filter by testID}

\item{quiet}{Suppress messages (default: \code{FALSE})}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
The object returned depends on the data available and any filters
(e.g. the argument  \code{modelID}):

If no filters are specified, then the object returned is a nested list,
similar to the original input object.

If one or more filters are applied, then only a subset of the original nested
list is returned. Technically, each filter applied is used to subset the
original nested list using \code{\link[base:Extract]{[[}}.

If no data is available, an empty list is returned (\code{list()}).
}
\description{
Get tests from a study
}
