% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getUpsetCols}
\alias{getUpsetCols}
\title{getUpsetCols}
\usage{
getUpsetCols(study, modelID, ...)
}
\value{
character vector
}
\description{
Determine the common columns across all tests of a model that are available
for filtering with UpSet.
}
