% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{getUpsetCols}
\alias{getUpsetCols}
\title{getUpsetCols}
\usage{
getUpsetCols(study, modelID)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}
}
\value{
character vector
}
\description{
Determine the common columns across all tests of a model that are available
for filtering with UpSet.
}
