% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importStudy}
\alias{importStudy}
\title{Import a study package}
\usage{
importStudy(study, libraries = NULL)
}
\arguments{
\item{study}{Name of an installed OmicNavigator study}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
Returns the \code{onStudy} object imported from the OmicNavigator
study package
}
\description{
Create an onStudy object by importing an installed study package
}
