% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{installStudy}
\alias{installStudy}
\title{Install a study as an R package}
\usage{
installStudy(study, library = .libPaths()[1])
}
\arguments{
\item{study}{An OmicAnalyzer study to install (class \code{oaStudy})}

\item{library}{Directory to install package. Defaults to first directory
returned by \code{\link{.libPaths}}.}
}
\description{
Install a study as an R package
}
