% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lane}
\alias{lane}
\title{lane from Bioconductor workflow RNAseq123}
\format{
A factor with 3 levels:
\describe{
\item{L004}{Sample sequenced on lane 4}
\item{L006}{Sample sequenced on lane 6}
\item{L008}{Sample sequenced on lane 8}
}
}
\source{
\url{https://bioconductor.org/packages/release/workflows/vignettes/RNAseq123/inst/doc/limmaWorkflow.html}
}
\usage{
lane
}
\description{
A subset of the object \code{lane} from Bioconductor workflow RNAseq123.
}
\examples{
  table(lane)
  str(lane)

}
\references{
Law CW, Alhamdoosh M, Su S, Dong X, Tian L, Smyth GK, Ritchie ME.
\href{https://f1000research.com/articles/5-1408/v3}{RNA-seq analysis is easy as 1-2-3 with limma, Glimma and edgeR \link{version 3; peer review: 3 approved}.}
F1000Research 2018, 5:1408
\doi{10.12688/f1000research.9005.3}

Sheridan, J.M., Ritchie, M.E., Best, S.A. et al.
\href{https://bmccancer.biomedcentral.com/articles/10.1186/s12885-015-1187-z}{A pooled shRNA screen for regulators of primary mammary stem and progenitor cells identifies roles for \emph{Asap1} and \emph{Prox1}.}
BMC Cancer 2015, 15:221
\doi{10.1186/s12885-015-1187-z}
}
\keyword{datasets}
