% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{listStudies}
\alias{listStudies}
\title{List available studies and their metadata}
\usage{
listStudies(libraries = NULL)
}
\arguments{
\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\value{
Returns a nested list with one element per installed OmicNavigator
study package. Each study package entry has the following sublist components:

\item{name}{(character) Name of the study}
\item{package}{(list) The fields from \code{DESCRIPTION}}
\item{results}{(nested list) The testIDs available for each modelID}
\item{enrichments}{(nested list) The annotationIDs available for each modelID}
}
\description{
List available studies and their metadata
}
