% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{OmicNavigator-package}
\alias{OmicNavigator}
\alias{OmicNavigator-package}
\title{OmicNavigator}
\description{
Package options to control package-wide behavior are described below.
}
\details{
The default prefix for OmicNavigator study packages is "ONstudy". If you
would prefer to use a different prefix, you can change the package option
\code{OmicNavigator.prefix}. For example, to use the prefix "OmicNavigatorStudy",
you could add the following line to your \code{.Rprofile} file.

\preformatted{
options(OmicNavigator.prefix = "OmicNavigatorStudy")
}

By default OmicNavigator will attempt to use the command-line tool \code{jq}
to obtain the description of each annotationID when the app calls
\code{\link{getEnrichmentsAnnotations}}. If this causes a problem, you can
disable this behavior via the package option
\code{OmicNavigator.useJqIfAvailable}. To do so, add the following line to
your \code{.Rprofile} file.

\preformatted{
options(OmicNavigator.useJqIfAvailable = FALSE)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abbvie-external/OmicNavigator}
  \item Report bugs at \url{https://github.com/abbvie-external/OmicNavigator/issues}
}

}
\author{
\strong{Maintainer}: Brett Engelmann \email{brett.engelmann@abbvie.com}

Authors:
\itemize{
  \item Terrence Ernst (Web application)
  \item John Blischak \email{jdblischak@gmail.com} (\href{https://orcid.org/0000-0003-2634-9879}{ORCID})
  \item Paul Nordlund (Web application)
  \item Justin Moore (UpSet-related functions and web application)
  \item Joe Dalen (Barcode functionality and web application)
  \item Akshay Bhamidipati (Web application)
  \item Marco Curado (Improved plotting capabilities)
  \item Joe LoGrasso (Support for plotly)
}

Other contributors:
\itemize{
  \item Elyse Geoffroy [contributor]
  \item AbbVie Inc. [copyright holder, funder]
}

}
