% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStudy}
\alias{plotStudy}
\title{Plot a feature using a custom plotting function}
\usage{
plotStudy(study, modelID, featureID, plotID, libraries = NULL)
}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{featureID}{Filter by featureID}

\item{plotID}{Filter by plotID}

\item{libraries}{The directories to search for installed study packages. If
left as \code{NULL} (the default), then
\code{\link[utils]{installed.packages}} will use the result of
\code{\link{.libPaths}}.}
}
\value{
This function is called for the side effect of creating a plot.
  However, it also invisible returns the original \code{onStudy} object
  passed to \code{study}.
}
\description{
Plot a feature using a custom plotting function
}
\seealso{
\code{\link{addPlots}}, \code{\link{getPlottingData}}
}
