% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStudy}
\alias{plotStudy}
\title{Plot a feature using a custom plotting function}
\usage{
plotStudy(study, modelID, featureID, plotID)
}
\arguments{
\item{study}{An OmicAnalyzer study. Either an object of class \code{oaStudy},
a \code{\link[RSQLite]{SQLiteConnection-class}} to a study database, or the name
of an installed study package.}

\item{modelID}{Filter by modelID}

\item{featureID}{Filter by featureID}

\item{plotID}{Filter by plotID}
}
\description{
Plot a feature using a custom plotting function
}
