% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{shared-add}
\alias{shared-add}
\title{Shared parameters for add functions}
\arguments{
\item{study}{An OmicAnalyzer study created with \code{\link{createStudy}}}

\item{overwrite}{Overwrite existing value}
}
\description{
Shared parameters for add functions
}
