% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{shared-add}
\alias{shared-add}
\title{Shared parameters for add functions}
\arguments{
\item{study}{An OmicNavigator study created with \code{\link{createStudy}}}

\item{reset}{Reset the data prior to adding the new data (default:
\code{FALSE}). The default is to add to or modify any previously added data
(if it exists). Setting \code{reset = TRUE} enables you to remove existing
data you no longer want to include in the study.}
}
\value{
Returns the original \code{onStudy} object passed to the argument
\code{study}, but modified to include the newly added data
}
\description{
Shared parameters for add functions
}
\keyword{internal}
