% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{shared-get}
\alias{shared-get}
\title{Shared parameters for get functions}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{testID}{Filter by testID}

\item{annotationID}{Filter by annotation}

\item{termID}{Filter by termID}

\item{featureID}{Filter by featureID}

\item{plotID}{Filter by plotID}
}
\description{
Shared parameters for get functions
}
\keyword{internal}
