% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{shared-get}
\alias{shared-get}
\title{Shared parameters for get functions}
\arguments{
\item{study}{An OmicNavigator study. Either an object of class \code{onStudy},
or the name of an installed study package.}

\item{modelID}{Filter by modelID}

\item{testID}{Filter by testID}

\item{annotationID}{Filter by annotationID}

\item{termID}{Filter by termID}

\item{featureID}{Filter by featureID}

\item{plotID}{Filter by plotID}

\item{quiet}{Suppress messages (default: \code{FALSE})}

\item{libraries}{Character vector of library directories to search for study
packages. If \code{NULL}, uses \code{.libPaths}.}
}
\value{
The object returned depends on the data available and any filters
(e.g. the argument  \code{modelID}):

If no filters are specified, then the object returned is a nested list,
similar to the original input object.

If one or more filters are applied, then only a subset of the original nested
list is returned. Technically, each filter applied is used to subset the
original nested list using \code{\link[base:Extract]{[[}}.

If no data is available, an empty list is returned (\code{list()}).
}
\description{
Shared parameters for get functions
}
\keyword{internal}
