% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{shared-upset}
\alias{shared-upset}
\title{Shared parameters for upset functions}
\arguments{
\item{anchor}{The primary test to filter from.}

\item{mustTests}{The tests whose significant values must be included. (The intersection)}

\item{notTests}{The tests whose significant values will be removed. (The difference)}

\item{sigValue}{The significance levels for each column.}

\item{operator}{The operators for each column.}

\item{column}{The columns to be thresheld.}

\item{type}{Type of p-value (\code{"nominal"} or \code{"adjusted"})}

\item{tests}{Restrict UpSet plot to these tests}

\item{legacy}{Use legacy code (for testing purposes only)}
}
\description{
Shared parameters for upset functions
}
\keyword{internal}
