% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{startApp}
\alias{startApp}
\title{Start app on local machine}
\usage{
startApp(...)
}
\arguments{
\item{...}{extra parameters passed to \link[opencpu]{ocpu_start_server}}
}
\description{
Start app on local machine
}
